make some struct lets vars instead
I am of the opinion that everything should be a `let` unless you ever
have to change it, however I consider `struct` properties to be a
special exception since this gives flexibility to the consumer.

i.e. if you do `let authToken = OAuth2Token(...)`, you won't be able to
change the vars anyway because the whole value is a constant.